+++ angular-workspace/projects/example-client-app/src/app/customapp/customapp.component.ts
@@ -21,6 +21,7 @@ interface SimpleTableRecord extends TableRecord {
     statusCode: number;
     result: string;
     number: number;
+    duration: number;
 }
 
 @Component({
@@ -136,7 +137,8 @@ export class CustomAppComponent {
                 date: (tableData.length % 2 === 0) ? new Date(2023, 7, 16, 3, 56, 11).valueOf() : new Date(2022, 2, 7, 20, 28, 41).valueOf(),
                 statusCode: (tableData.length % 2 === 0) ? 100 : 101,
                 result: (tableData.length % 2 === 0) ? 'success' : 'unknown',
-                number: tableData.length / 10
+                number: tableData.length / 10,
+                duration: tableData.length * 1000 * (1.1 + 2 * 60 + 3 * 3600)
             });
         }
         this.tableDataSubject.next(tableData);