+++ packages/nimble-components/src/icon-base/tests/icon-metadata.ts
@@ -31,8 +31,14 @@ export const iconMetadata: {
     IconArrowExpanderUp: {
         tags: []
     },
+    IconArrowInCircle: {
+        tags: ['move in']
+    },
     IconArrowLeftFromLine: {
         tags: ['logout']
+    },
+    IconArrowOutCircle: {
+        tags: ['move out']
     },
     IconArrowPartialRotateLeft: {
         tags: ['reset']
@@ -57,6 +63,9 @@ export const iconMetadata: {
     },
     IconArrowUp: {
         tags: ['sort', 'arrow', 'up', 'ascending']
+    },
+    IconAt: {
+        tags: ['@', 'email', 'handle', 'mention']
     },
     IconBars: {
         tags: ['hamburger menu']
@@ -250,6 +259,9 @@ export const iconMetadata: {
     IconFloppyDiskCheckmark: {
         tags: ['save', 'no unsaved changed']
     },
+    IconFloppyDiskPen: {
+        tags: ['save as', 'pencil']
+    },
     IconFloppyDiskStarArrowRight: {
         tags: ['save', 'autosave']
     },
@@ -352,6 +364,9 @@ export const iconMetadata: {
     IconMobile: {
         tags: ['phone']
     },
+    IconNi: {
+        tags: ['National Instruments', 'logo', 'product']
+    },
     IconNotebook: {
         tags: []
     },
@@ -405,6 +420,9 @@ export const iconMetadata: {
     },
     IconSquareT: {
         tags: ['static text']
+    },
+    IconSystemlink: {
+        tags: ['logo', 'product']
     },
     IconT: {
         tags: ['text']